feat: create PrivateRoute
