--- src/firebase/firebase.ts
@@ -1,92 +1,0 @@
-import { initializeApp } from 'firebase/app';
-import {
-  GoogleAuthProvider,
-  getAuth,
-  signInWithPopup,
-  signInWithEmailAndPassword,
-  createUserWithEmailAndPassword,
-  sendPasswordResetEmail,
-  signOut,
-} from 'firebase/auth';
-import {
-  getFirestore, query, getDocs, collection, where, addDoc
-} from 'firebase/firestore';
-
-const firebaseConfig = {
-  apiKey: 'AIzaSyAnTTxWvKYAbvoz0kdUXrhVK3TkmVTeFG8',
-  authDomain: 'graphiql-aut-db.firebaseapp.com',
-  projectId: 'graphiql-aut-db',
-  storageBucket: 'graphiql-aut-db.appspot.com',
-  messagingSenderId: '1062970047611',
-  appId: '1:1062970047611:web:e35fe9e716175f640076de',
-};
-
-const app = initializeApp(firebaseConfig);
-const auth = getAuth(app);
-const db = getFirestore(app);
-
-const googleProvider = new GoogleAuthProvider();
-
-const signInWithGoogle = async () => {
-  try {
-    const res = await signInWithPopup(auth, googleProvider);
-    const { user } = res;
-    const q = query(collection(db, 'users'), where('uid', '==', user.uid));
-    const docs = await getDocs(q);
-    if (docs.docs.length === 0) {
-      await addDoc(collection(db, 'users'), {
-        uid: user.uid,
-        name: user.displayName,
-        authProvider: 'google',
-        email: user.email,
-      });
-    }
-  } catch (err) {
-    console.error(err);
-  }
-};
-
-const logInWithEmailAndPassword = async (email: string, password: string) => {
-  try {
-    await signInWithEmailAndPassword(auth, email, password);
-  } catch (err) {
-    console.error(err);
-  }
-};
-
-const registerWithEmailAndPassword = async (name: string, email: string, password: string) => {
-  try {
-    const res = await createUserWithEmailAndPassword(auth, email, password);
-    const { user } = res;
-    await addDoc(collection(db, 'users'), {
-      uid: user.uid,
-      name,
-      authProvider: 'local',
-      email,
-    });
-  } catch (err) {
-    console.error(err);
-  }
-};
-
-const sendPasswordReset = async (email: string) => {
-  try {
-    await sendPasswordResetEmail(auth, email);
-  } catch (err) {
-    console.error(err);
-  }
-};
-
-const logout = () => {
-  signOut(auth);
-};
-
-export {
-  auth,
-  db,
-  signInWithGoogle,
-  logInWithEmailAndPassword,
-  registerWithEmailAndPassword,
-  sendPasswordReset,
-  logout,
-};
